move gamification models withing common
