+++ libs/server/common/src/core/models/index.ts
@@ -1,4 +1,3 @@
 export * from './common.model';
-export * from './gamification-domain.model';
 export * from './github-shared.model';
