+++ libs/server/common/src/gamification/index.ts
@@ -1,4 +1,5 @@
+export * from './gamification-domain.model';
 export * from './levels/level.service';
 export * from './levels/level.utils';
-export * from './xp/xp-calculator.service';
 export * from './xp-database.service';
+export * from './xp/xp-calculator.service';
